Complete Phase 8: Feature Enhancements
- Enhanced array views with zero-copy using Arc/Weak reference counting
- Implemented true memory-mapped arrays using memmap2 crate
- Added robust reference counting system with Arc/Weak for views
- Enhanced API coverage: copy(), as_contiguous(), atleast_*d(), moveaxis()
- Fixed all Clippy warnings and improved code quality
- Added comprehensive test coverage (264 tests total, 27 new tests)
- Updated roadmap to reflect Phase 8 completion

All tests passing, all linting warnings resolved.

+++ src/indexing/advanced/fancy.rs
@@ -34,10 +34,10 @@ pub fn fancy_index_array(array: &Array, indices: &Array) -> Result<Array, IndexE
                 let mut idx = *index_ptr.add(i) as i64;
                 // Normalize negative indices
                 if idx < 0 {
-                    idx += array_shape[0] as i64;
+                    idx += array_shape[0];
                 }
                 // Validate bounds
-                if idx < 0 || idx >= array_shape[0] as i64 {
+                if idx < 0 || idx >= array_shape[0] {
                     return Err(IndexError::OutOfBounds);
                 }
                 Ok(idx)
